Now ego supports float values

+++ src/ast/nodes/number.rs
@@ -1,17 +1,17 @@
 #[derive(Debug, Clone)]
 pub struct Number {
-    pub value: i64,
+    pub value: f64,
     pub at: usize,
     pub line: usize,
 }
 
 impl Number {
-    pub fn new(value: i64, at: usize, line: usize) -> Number {
+    pub fn new(value: f64, at: usize, line: usize) -> Number {
         Number { value, at, line }
     }
 
     pub fn from_string(value: String, at: usize, line: usize) -> Option<Number> {
-        value.parse::<i64>().ok().map(|number| Number {
+        value.parse::<f64>().ok().map(|number| Number {
             value: number,
             at,
             line,
